+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Parable Console
 
+## 0.2.0
+
+_Changes_
+
+- `App` has been renamed to `Application` for consistency reasons.
+- Running a `Command` (or the `Application`) will no longer return any values. All are typed to return `void`.
+
 ## 0.1.3
 
 _Changes_